Rename index to icons
