+++ src/components/icons/Icon.tsx
@@ -1,5 +1,5 @@
 import { createIconComponent, glyphs } from "@leafygreen-ui/icon";
-import * as icons from "./index";
+import * as icons from "./icons";
 
 const glyphMap = {
   ...glyphs,
